Move ServiceMain into its own .cc/.h
